Move `MockedProvider` into new common testing utility location
